Scrape Yelp for Entire Cities
The code is quite simple but relies on a crucial aspect: a grid of
points over your city. Yelp's API allows one to search within a radius
of a point so we created a dense grid of points for San Diego and
collected most of the businesses Yelp has on its site.

Once the business URLs are acquired, we just loop through them, scraping
its priceRange, number of reviews, rating, and most importantly, all of
the reviews themselves(including text).

Given the above data, we then extract all of the unique users and scrape
all of their data, including Location, Elite status, number of friends,
number of reviews, and most importantly, friend IDs and review dates.